+++ src/error.c
@@ -1,4 +1,4 @@
 #include <stdlib.h>
 #include <sea-turtle.h>
 
-_Thread_local size_t sea_turtle_error = SEA_TURTLE_ERROR_NONE;
+_Thread_local uintmax_t sea_turtle_error = SEA_TURTLE_ERROR_NONE;
